a few type signature fix and adding the skeleton of interactive voxelpainter example

+++ threejs/src/main/kotlin/three/math/Math.kt
@@ -5,7 +5,7 @@ package three.math
 @JsName("Math")
 external class Math {
     companion object {
-        fun randFloatSpread(int: Int): Double
+        fun randFloatSpread(int: Float): Float
     }
 
 }
